Implemented options
Created the class OptionCriticNetwork in PPO torch. Actor consists of state feature extractor MLP then layers to obtain inter-option policy prob distribution, option termination probabilities, and paramenter for the weight and biases of the intra-option policies. Separate MLP for critic. Actor and Critic loss are computed the same way as in the PPO agent.